+++ scripts/io/ISettingsManagerAsync.ts
@@ -1,6 +1,8 @@
 interface ISettingsManagerAsync {
     getValueAsync<T>(key:string, fallback?:T):Promise<T>;
     setValueAsync<T>(key:string, value:T):Promise<void>;
+    removeValueAsync(key:string):Promise<void>;
+    clearAsync():Promise<void>;
 }
 
 export default ISettingsManagerAsync