+++ scripts/net/HttpClient.ts
@@ -4,7 +4,7 @@ import "whatwg-fetch";
 import HttpResponse from "./HttpResponse";
 import {injectable} from "inversify";
 import Dictionary from "../util/Dictionary";
-import * as _ from "lodash";
+import {merge} from "lodash";
 
 
 const binaryMimeTypes = ["application/octet-stream", "application/pdf", "application/zip", "application/x-", "image/", "video/"];
@@ -16,12 +16,12 @@ class HttpClient implements IHttpClient {
         return this.performNetworkCall(url, "get", undefined, headers);
     }
 
-    post(url: string, body: {} | FormData, headers?: Dictionary<string>): Observable<HttpResponse> {
-        return this.performNetworkCall(url, "post", this.getJsonBody(body), this.addJsonHeaders(headers));
+    post(url: string, body: {}|FormData, headers?: Dictionary<string>): Observable<HttpResponse> {
+        return this.performNetworkCall(url, "post", this.getJsonBody(body), this.addJsonHeaders(headers, body));
     }
 
     put(url: string, body: {}, headers?: Dictionary<string>): Observable<HttpResponse> {
-        return this.performNetworkCall(url, "put", this.getJsonBody(body), this.addJsonHeaders(headers));
+        return this.performNetworkCall(url, "put", this.getJsonBody(body), this.addJsonHeaders(headers, body));
     }
 
     delete(url: string, headers?: Dictionary<string>): Observable<HttpResponse> {
@@ -29,11 +29,13 @@ class HttpClient implements IHttpClient {
     }
 
     private getJsonBody(body: {} | FormData) {
-        return !(body instanceof FormData) ? JSON.stringify(body) : body;
+        return !(this.isFormData(body)) ? JSON.stringify(body) : body;
     }
 
-    private addJsonHeaders(headers: Dictionary<string>) {
-        return _.merge({}, {
+    private addJsonHeaders(headers: Dictionary<string>, body: {} | FormData) {
+        return this.isFormData(body)
+            ? headers
+            : merge({}, {
             "Accept": "application/json",
             "Content-Type": "application/json"
         }, headers);
@@ -64,6 +66,10 @@ class HttpClient implements IHttpClient {
 
     private isBinaryPayload(contentType: string): boolean {
         return binaryMimeTypes.some(type => !!contentType.match(type));
+    }
+
+    private isFormData(data: any): boolean {
+        return data instanceof FormData;
     }
 }